Refactor - extract DefaultBootstrapDiscoveryContext to the package level after the review.
